Remove backup functionality
The backup functionality and startup script is no longer needed
since the functionality has been removed in the WAL-E container.

This also removes a dependency on a particular backup solution,
and keeps the image closer to the PostgreSQL default.

Backups should instead be handled as appropriate for the particular
runtime environment. For instance, with Kubernetes, one can use
init containers for restoring backups.