Add client tools and run timescaledb-tune
This adds timescaledb-tune and timescaledb-parallel-copy to the
Docker image for easier usage of these tools for Docker users. In
addition, timescaledb-tune is now run on the initial run of the
Docker image so that the user does not have to manually invoke it.
This should make Docker setup very easy.