Add build targets for OSS-only images
For users who want to continue using TimescaleDB, but only want
to use completely open-source bits (i.e., no TSL code), we add
build targets to generate images for that. Additionally the
Makefile is simplified by defining a new PG_VER_NUMBER that can
be used to pull the right up-to-date image for each PostgreSQL
major version. In the future we should remove PG_VER and simplify
the Makefile further.