Pass docker resource limits to timescaledb-tune by default
This fixes an issue where the system calls used by timescaledb-tune were
reporting the resources available on a container's host, which often
do not match the resources allocated to a given container. To fix this,
we inspect the cgroups metadata to determine the actual resource limits
imposed on the container. We also allow users to override these values.