Use free -m for memory to prevent overflow issues
Using free -b caused the total memory to run into the free memory,
thus creating a gigantic number as the "free memory." This would
then cause the values returned by timescaledb-tune to be way off
and prevent PostgreSQL from starting. Instead we now use megabytes
and then multiple them by 2^20 to get the (rough) amount of bytes
on the system.