Skip memory setting for 64-bit unsigned max
Some Docker containers may report the memory limit as being equal
to the 64-bit unsigned max, which bash cannot handle very well. We
add a simple check to see if the value is that special value, and
if so we just let timescaledb-tune try to discover the limit. The
big value may cause either bash to crash or for the limit passed
to timescaledb-tune to be negative (which results in a panic).