Updated upstream Bitnami image versions and added fixes to match them
The previous images did not provide the correct POSTGRESQL_DATA_DIR environment variable, causing the Timescale init scripts to fail.
The init scripts also failed to work properly with the -nounset (-u) script flag which was introduced in the upstream image.
The first init script was fixed to avoid a warning due to the extension creation being attempted twice.
The Dockerfile was fixed to preload the extension (as the main image does) -- an omission of which prevented the init scripts from working at all.

Fixes #50, fixes #56.