Update Dockerfiles to use explicit Go version
Using the most up-to-date version of Go is important so that our
tools don't have any security issues. By making the Go version an
argument to base image, we can ensure we are not using cached older
versions.

Also this cleans up some other bits of the Dockerfiles, including
reducing the number of older images stored in our main image.