Delete Telemetry job when telemetry is disabled
If Telemetry is disabled by setting `TIMESCALEDB_TELEMETRY` to `off`,
the job is still present in the `bgw_job` table and will still be
processed as a job and generate errors and warnings even though it does
not do anything when executed.

This commit fixes this by deleting the Telemetry job during
configuration if the `TIMESCALEDB_TELEMETRY` is set to `off`.