Remove hardcoded versions from Dockerfiles
This patch removes the hardcoded versions from the Dockerfile and
changes them into arguments. This removes the need to update the
version in multiple places when a new release is done and it also
removes the need to maintain the previous version as that can be
looked up. With this patch only 1 place in the CI file needs to be
updated on a new release.