Make memory tuning cgroups v2 compatible
The old implementation only checked the sysfs files corresponding to
cgroups v1. Most Linux distribution kernels have switched to cgroups
v2 several years ago, which uses a different set of paths for various
settings (memory limits included).

The net effect of the old implementation was that it completely
ignored the memory limits, if the container was run in a kernel with
cgroups v2 (in those kernels cgroups v1 are disabled by default).